+++ src/routes.js
@@ -1,11 +1,13 @@
 import React from "react";
-import { Route } from "react-router";
+import { Route, IndexRoute } from "react-router";
 
 // Components
 import App from "./components/app";
-//import Home from "./views/home";
+import Home from "./views/home";
 //import Docs from "./views/docs";
 
 module.exports = (
-    <Route path="/" component={App} />
+  <Route path="/" component={App}>
+    <IndexRoute component={Home} />
+  </Route>
 );
